# backend/api/swagger.yaml
openapi: "3.0.0"
info:
  title: API Music Room Suricates 
  description: This is a MusicRoom server with all the requests available to the users.
    You can authentify yourself to get access to the requests.
  version: 1.0.0
servers:
  - url: "http://62.34.5.191:45559/api"
  - url: "http://localhost:45559/api"
  - url: "http://localhost:46000/api"
...
paths:
...
  /me:
    get:
      tags:
        - "user"
      summary: "Get your user profile"
      operationId: "getMyProfile"
      responses:
        200:
          description: |
            __{code:__ '0'__, user:__ *UserObject*__}__ - `User profile successfully returned!`
          content:
           application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: "0"
                user:
                  $ref: '#/components/schemas/User'
                  example:
                    $ref: '#/components/examples/userExample'
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
...
  /users/{userId}/attachGoogleAccount:
    post:
      tags:
        - "user"
      summary: "Attach the user's google account"
      operationId: "attachgoogleAccount"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of the user"
        required: true
        schema:
          type: "string"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                token:
                  type: string
                MusicRoom_ID:
                  type: string
              required:
                - token
                - MusicRoom_ID
      responses:
        200:
          description: |
            __{code:__ '0'__, user:__ *userObject*__}__ - `Google account has been successfully attached!`
          content:
           application/json: 
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: "0"
                user:  
                  $ref: '#/components/schemas/User'
                  example:
                    $ref: '#/components/examples/userExample'
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Missing token or MusicRoom_ID field!`

            __{code:__ '1'__, message:__ *errorMessage*__}__ - `Auth with Google failed, invalid token or client id!`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500:
          description: |
            __{code:__ '0'__}__ - `Internal error!`
  /users/{userId}/password:
    post:
      tags:
      - "user"
      summary: "Changes the password of the specified user"
      operationId: "resetPassword"
      parameters:
        - name: "userId"
          in: "path"
          description: "ID of the user"
          required: true
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  description: "User current password"
                newPassword:
                  type: string
                  description: "User new password - __Min 5 char - Authorized char: [a-zA-Z0-9éèàêô,'!@# -]__"
              required:
                - currentPassword
                - newPassword
      responses:
        200: 
          description: "Password successfully updated!"
        400:
          description: |
            __{code:__ '0'__, message:__ *errorMessage*__}__ - `Invalid current password!`

            __{code:__ '1'__, message:__ *errorMessage*__}__ - `Invalid new password format!`
        401:
          description: |
            __{error:__ *authError*__}__ - `Authentification failed, invalid token or userId!`
        500: 
          description: "Error - Internal server error. Error object returned."
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        userInfo:
          type: object
          properties:
            _id:
              type: string
            active:
              type: boolean
            email:
              type: string
              format: email
            password:
              type: string
              format: password
            firstName:
              type: string
            lastName:
              type: string
            pseudo:
              type: string
            registrationData:
              type: string
              format: date-time
            musicalPreferences:
              type: array
              items:
                type: string
        userData:
          type: object
          properties:
            _id:
              type: string
            playlists:
              type: array
              items:
                type: object
                properties:
                  playlist:
                    type: string
                  playlistType:
                    type: number
            friendsId:
              type: array
              items:
                type: string
            events:
              type: array
              items:
                type: object
                properties:
                  eventsId:
                    type: string
                  songIdVotes:
                    type: array
                    items:
                      type: string
...
  examples:
...
    userExample:
      _id: 60104b6b5bb03207e06857d5
      userInfo:
        _id: 601014ae2fba2f1bbc9354fc
        active: true
        email: nomemailoriginale@musicroom.io
        password: $2b$10$iXFmPaynUUqdUiRd6B3RW.k8deKJMQO0Tr.CZExmC30t/O8AxfBrC
        firstName: Jean-Claude
        lastName: Van Damme
        pseudo: JCVD
        registrationDate: 2021-01-25T15:10:32.148+00:00
        musicalPreferences: 
          - salsa
          - rap
          - soul
      userData:
        playlists:
          - playlist: a2f1bbc9354fc60104b6b5bb03
            playlistType: 0
          - playlist: fc60104b6b5bb03a2f1bbc9354
            playlistType: 2
          - playlist: 04b6b5bbfc60103a2f1bbc9354
            playlistType: 1
        friendsId:
          - 9354fc6010a2f1bbc93b6b5b
          - 9356010a93b6010a4fc6010a
          - 60104b6b5bb03207e06857d5
        events:
          - eventsId: 03207e06857d59354fc6010a2
            songIdVotes:
            - fc6010aa2f1bbc9354fc60a201
            - aa2f1bbc9354fc60afc6010201
          - eventsId: 7d59354fc6010a203207e0685
            songIdVotes:
            - bc9354fc60a20135fc10201a0
            - fc6010aa2f1baa2f1bbc960f4
            - 7d59354fc6010a203207e06d2
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []